some abstractions
