+++ src/packages/signMessage.ts
@@ -1,0 +1,1 @@
+export const signMessage = async () => {};
