+++ src/types.ts
@@ -1,0 +1,9 @@
+export interface KeypairBase64 {
+  publicKey: string;
+  privateKey: string;
+}
+
+export interface KeypairArrayBuffer {
+  publicKey: ArrayBuffer;
+  privateKey: ArrayBuffer;
+}
